Refactored list sessions
