+++ Maestro/src/app/list-session.html
@@ -1,0 +1,21 @@
+<ion-list>
+    <ion-item-sliding *ngFor="let w of sessions" #slidingItem>
+        <ion-item>
+            <div>
+                <h2>{{w.title}}</h2>
+                <h3>{{w.session_duration}}</h3>
+            </div>
+        </ion-item>
+        <ion-item-options>
+            <button ion-button (click)="sessionInfo(w, slidingItem)" color="primary" icon-left>
+
+                <ion-icon id="more" name="ios-more"></ion-icon>
+                More
+            </button>
+            <button ion-button (click)="deletePrompt()" color="danger">
+                <ion-icon id="delete" name="close-circle"></ion-icon>
+                Delete
+            </button>
+        </ion-item-options>
+    </ion-item-sliding>
+</ion-list>
