+++ Maestro/src/pages/session-hist/session-hist.html
@@ -11,25 +11,5 @@
 
 
 <ion-content padding>
-    <ion-list>
-        <ion-item-sliding *ngFor="let w of sessions" #slidingItem >
-            <ion-item >
-                <div>
-                    <h2>{{w.title}}</h2>
-                    <h3>{{w.session_duration}}</h3>
-                </div>
-            </ion-item>
-            <ion-item-options>
-                <button ion-button  (click)="sessionInfo(w, slidingItem)" color="primary" icon-left>
-
-                    <ion-icon id="more" name="ios-more"></ion-icon>
-                    More
-                </button>
-                <button ion-button  (click)="deletePrompt()" color="danger">
-                    <ion-icon id="delete" name="close-circle"></ion-icon>
-                    Delete
-                </button>
-            </ion-item-options>
-        </ion-item-sliding>
-    </ion-list>
+    <list-session [amount]="10000"></list-session>
 </ion-content>
